Attempt to remove various warnings under MSVC 2003, mostly related to casting, but also some instances of inaccessible code.
git-svn-id: svn://svn.icculus.org/fs2open/trunk/fs2_open@3898 387891d4-d844-0410-90c0-e4c51a9137d3

+++ code/cutscene/oggplayer.cpp
@@ -1,12 +1,17 @@
 /*
  * $Logfile: /Freespace2/code/cutscene/oggplayer.cpp $
- * $Revision: 2.1 $
- * $Date: 2007-01-07 12:29:43 $
- * $Author: taylor $
+ * $Revision: 2.2 $
+ * $Date: 2007-01-14 10:26:37 $
+ * $Author: wmcoolmon $
  *
  * movie player code
  * 
  * $Log: not supported by cvs2svn $
+ * Revision 2.1  2007/01/07 12:29:43  taylor
+ * add Theora player
+ * remove DirectShow support from movie player
+ * fix triple-buffer page flipping problem (Mantis bug #1190)
+ *
  * Revision 1.0  2006/12/25 21:44:11  taylor
  *
  * $NoKeywords: $
@@ -875,7 +880,7 @@ void theora_play(THEORAFILE *movie)
 						if (val < -32768)
 							val = -32768;
 
-						audiobuf[count++] = val;
+						audiobuf[count++] = (short)val;
 					}
 				}
